refactor: timer feature를 timer-log로 병합
