+++ client/src/entities/timer-log/index.ts
@@ -1,3 +1,6 @@
 export * from "./model/types";
+export * from "./model/reducers";
+export * from "./lib/selectors";
 export * from "./lib/parse-dto";
+export * from "./ui/formatters";
 export * from "./api/types";
