+++ client/src/entities/timer-log/model/types.ts
@@ -5,3 +5,10 @@ export interface TimerLog {
   type: "start" | "stop" | "add" | "sub";
   createdAt: Date;
 }
+
+export interface TimerState {
+  initialMs: number; // 초기 시간 (ms)
+  offsetMs: number; // 증감된 시간 (ms)
+  accumulatedMs: number; // 타이머가 정지할 때까지의 누적 시간 (ms), 타이머가 정지될 때마다 갱신됨
+  startedAt: number | null; // 타이머 시작 시간 (ms), null이면 타이머가 정지되어 있음
+}
