+++ client/src/entities/timer-log/ui/formatters.ts
